# internal/server/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Bongsu Vulnerability Management API
  version: 0.1.0
  description: >
    Vulnerability management and SBOM platform API. Provides host inventory,
    vulnerability scanning, CVE database management, RBAC, and installer
    endpoints.
  license:
    name: MIT
servers:
  - url: /
    description: Current server
security: []
paths:
...
  /api/auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      id: { type: string }
                      username: { type: string }
                      role: { type: string }
                  expires:
                    type: string
                    format: date-time
        "400":
          description: Missing username or password
        "401":
          description: Invalid credentials
  /api/auth/logout:
    post:
      operationId: authLogout
      tags: [auth]
      summary: Logout and clear session
      security:
        - apiKey: []
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
  /api/auth/me:
    get:
      operationId: authMe
      tags: [auth]
      summary: Get current user info
      security:
        - apiKey: []
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id: { type: string }
                      username: { type: string }
                      role: { type: string }
        "401":
          description: Not authenticated
  /api/auth/change-password:
    post:
      operationId: authChangePassword
      tags: [auth]
      summary: Change password for logged-in user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 12
      responses:
        "200":
          description: Password changed
        "400":
          description: Missing or invalid input
        "401":
          description: Not authenticated or wrong current password

  # ── Health ──────────────────────────────────────────────────────────────
...
  /api/hosts:
    get:
      operationId: listHosts
      tags: [hosts]
      summary: List all hosts
      security:
        - AdminKey: []
        - ViewerKey: []
      parameters:
        - name: agent_status
          in: query
          schema:
            type: string
            enum: [online, stale, offline, unknown]
          description: Filter by agent status
        - name: inventory_status
          in: query
          schema:
            type: string
            enum: [healthy, degraded, stale, empty, none]
          description: Filter by inventory status
        - name: agent_version_state
          in: query
          schema:
            type: string
            enum: [current, outdated, unknown]
          description: Filter by agent version drift state
      responses:
        "200":
          description: List of hosts with vulnerability counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostWithVulns"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/hosts/{id}:
    get:
      operationId: getHost
      tags: [hosts]
      summary: Get host details
      security:
        - AdminKey: []
        - ViewerKey: []
      parameters:
        - $ref: "#/components/parameters/HostID"
      responses:
        "200":
          description: Host details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Host"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/hosts/{id}/packages:
    get:
      operationId: getHostPackages
      tags: [hosts]
      summary: Get host packages
      security:
        - AdminKey: []
        - ViewerKey: []
      parameters:
        - $ref: "#/components/parameters/HostID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paginated package list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPackages"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/hosts/{id}/sbom:
    get:
      operationId: getHostSBOM
      tags: [hosts]
      summary: Get host SBOM
      description: Export SBOM in CycloneDX or SPDX format.
      security:
        - AdminKey: []
        - ViewerKey: []
      parameters:
        - $ref: "#/components/parameters/HostID"
        - name: format
          in: query
          schema:
            type: string
            enum: [cyclonedx, cdx, spdx]
            default: cyclonedx
      responses:
        "200":
          description: SBOM document
          content:
            application/vnd.cyclonedx+json:
              schema:
                type: object
            application/spdx+json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Host not found or no packages available
  /api/hosts/{id}/vuln-counts:
    get:
      operationId: getHostVulnCounts
      tags: [hosts]
      summary: Get vulnerability counts per severity
      security:
        - AdminKey: []
        - ViewerKey: []
      parameters:
        - $ref: "#/components/parameters/HostID"
      responses:
        "200":
          description: Vulnerability counts keyed by severity
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/hosts/{id}/metadata:
    post:
      operationId: updateHostMetadata
      tags: [hosts]
      summary: Update host metadata
      security:
        - AdminKey: []
      parameters:
        - $ref: "#/components/parameters/HostID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostMetadataUpdate"
      responses:
        "200":
          description: Updated host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Host"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/hosts/{id}/agent-token/reset:
    post:
      operationId: resetHostAgentToken
      tags: [hosts]
      summary: Reset agent token
      security:
        - AdminKey: []
      parameters:
        - $ref: "#/components/parameters/HostID"
      responses:
        "200":
          description: Token reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Vulnerabilities ─────────────────────────────────────────────────────
  /api/vulnerabilities:
    get:
      operationId: listVulnerabilities
      tags: [vulnerabilities]
      summary: List vulnerabilities
      security:
        - AdminKey: []
        - ViewerKey: []
      parameters:
        - name: severity
          in: query
          schema:
            type: string
        - name: host_id
          in: query
          schema:
            type: string
        - name: triage_status
          in: query
          schema:
            type: string
            enum: [open, in_progress, accepted_risk, false_positive, fixed, ignored]
        - name: finding_source
          in: query
          schema:
            type: string
            enum: [scanner, cve-db]
        - name: risk_level
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
        - name: overdue
          in: query
          schema:
            type: boolean
        - name: exploited
          in: query
          schema:
            type: boolean
        - name: min_epss
          in: query
          schema:
            type: number
            format: double
        - name: min_epss_percentile
          in: query
          schema:
            type: number
            format: double
        - name: pkg_name
          in: query
          schema:
            type: string
        - name: container
          in: query
          schema:
            type: string
        - name: owner
          in: query
          schema:
            type: string
        - name: team
          in: query
          schema:
            type: string
        - name: environment
          in: query
          schema:
            type: string
        - name: criticality
          in: query
          schema:
            type: string
        - name: min_cvss
          in: query
          schema:
            type: number
            format: double
        - name: show_no_fix
          in: query
          schema:
            type: boolean
          description: Include vulnerabilities with no fix available
        - name: show_mismatch
          in: query
          schema:
            type: boolean
          description: Include mismatched vulnerabilities
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: sort_by
          in: query
          schema:
            type: string
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Paginated vulnerability list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedVulnerabilities"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "504":
          description: Vulnerability list timeout
  /api/vulnerabilities/export:
    get:
      operationId: exportVulnerabilities
      tags: [vulnerabilities]
      summary: Export vulnerabilities
      description: Export up to 100,000 rows as CSV or JSON.
      security:
        - AdminKey: []
        - ViewerKey: []
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, json]
            default: csv
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 100000
          description: Max rows to export
        - name: severity
          in: query
          schema:
            type: string
        - name: host_id
          in: query
          schema:
            type: string
        - name: triage_status
          in: query
          schema:
            type: string
        - name: finding_source
          in: query
          schema:
            type: string
        - name: risk_level
          in: query
          schema:
            type: string
        - name: pkg_name
          in: query
          schema:
            type: string
        - name: min_cvss
          in: query
          schema:
            type: number
            format: double
        - name: sort_by
          in: query
          schema:
            type: string
        - name: sort_order
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Exported file
          content:
            text/csv:
              schema:
                type: string
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/vulnerabilities/filters:
    get:
      operationId: getVulnFilters
      tags: [vulnerabilities]
      summary: Get available filter options
      security:
        - AdminKey: []
        - ViewerKey: []
      responses:
        "200":
          description: Filter options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilterOptions"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/vulnerabilities/triage:
    post:
      operationId: upsertVulnerabilityTriage
      tags: [vulnerabilities]
      summary: Upsert vulnerability triage
      security:
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VulnerabilityTriage"
      responses:
        "200":
          description: Triage upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VulnerabilityTriage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key

    bearerAuth:
      type: http
      scheme: bearer
      description: Session token

    AdminKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Admin API key

    AgentKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Agent API key (or admin key)

    ViewerKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Viewer API key (subject:key format)

    InstallToken:
      type: apiKey
      in: header
      name: X-Install-Token
      description: Install token

  # ── Reusable Parameters ──────────────────────────────────────────────
  parameters:
    HostID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Host ID
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        text/plain:
          schema:
            type: string
            example: unauthorized
...
  schemas:
...
    HostWithVulns:
      allOf:
        - $ref: "#/components/schemas/Host"
        - type: object
          properties:
            vuln_counts:
              type: object
              additionalProperties:
                type: integer
            active_vuln_counts:
              type: object
              additionalProperties:
                type: integer
            latest_inventory:
              $ref: "#/components/schemas/HostInventorySummary"
...